backend/src/tests/auth/perms.test.ts
import { describe, test } from 'mocha'
import { expect } from 'chai'
import { admin, user } from '../../api/auth/roles'
import { ensure_perms } from '../../api/auth/perms'

describe('Table perms', () => {
    test(ensure_perms.name, async () => {
        const result = await ensure_perms(
            { users: { id: true } },
            { user_id: 1, role_ids: [admin] },
            'query'
        )
        expect(result).to.equal(undefined)
    })
    test(ensure_perms.name, async () => {
        try {
            const result = await ensure_perms(
                { migrations: { id: true } },
                { user_id: 1, role_ids: [user] },
                'query'
            )
            expect(true).to.equal(false)
        } catch (error) {
            expect(error.message.length > 0).to.equal(true)
        }
    })
})
